Rename git/ssh -> git/gitssh
Avoid conflict with golang.org/x/crypto/ssh
